Revert "PF-1171: Add CLI command for BQ data table reference"
