+++ src/main/java/bio/terra/cli/businessobject/Resource.java
@@ -1,6 +1,5 @@
 package bio.terra.cli.businessobject;
 
-import bio.terra.cli.businessobject.resource.BqDataTable;
 import bio.terra.cli.businessobject.resource.BqDataset;
 import bio.terra.cli.businessobject.resource.GcpNotebook;
 import bio.terra.cli.businessobject.resource.GcsBucket;
@@ -55,7 +54,6 @@ public abstract class Resource {
   public enum Type {
     GCS_BUCKET,
     BQ_DATASET,
-    BQ_DATA_TABLE,
     AI_NOTEBOOK;
   }
 
@@ -106,8 +104,6 @@ public abstract class Resource {
         return new GcsBucket(wsmObject);
       case BIG_QUERY_DATASET:
         return new BqDataset(wsmObject);
-      case BIG_QUERY_DATA_TABLE:
-        return new BqDataTable(wsmObject);
       case AI_NOTEBOOK:
         return new GcpNotebook(wsmObject);
       default:
