+++ src/main/java/bio/terra/cli/command/resource/AddRef.java
@@ -1,6 +1,5 @@
 package bio.terra.cli.command.resource;
 
-import bio.terra.cli.command.resource.addref.BqDataTable;
 import bio.terra.cli.command.resource.addref.BqDataset;
 import bio.terra.cli.command.resource.addref.GcsBucket;
 import picocli.CommandLine;
@@ -14,5 +13,5 @@ import picocli.CommandLine;
     name = "add-ref",
     aliases = {"add-referenced"},
     description = "Add a new referenced resource.",
-    subcommands = {BqDataTable.class, BqDataset.class, GcsBucket.class})
+    subcommands = {BqDataset.class, GcsBucket.class})
 public class AddRef {}
