+++ src/main/java/bio/terra/cli/command/resource/addref/BqDataset.java
@@ -1,7 +1,6 @@
 package bio.terra.cli.command.resource.addref;
 
 import bio.terra.cli.command.shared.BaseCommand;
-import bio.terra.cli.command.shared.options.BigQueryIds;
 import bio.terra.cli.command.shared.options.Format;
 import bio.terra.cli.command.shared.options.ResourceCreation;
 import bio.terra.cli.command.shared.options.WorkspaceOverride;
@@ -18,7 +17,16 @@ import picocli.CommandLine;
     showDefaultValues = true)
 public class BqDataset extends BaseCommand {
   @CommandLine.Mixin ResourceCreation resourceCreationOptions;
-  @CommandLine.Mixin BigQueryIds bigQueryIds;
+
+  @CommandLine.Option(
+      names = "--project-id",
+      required = true,
+      description = "GCP project id of the dataset.")
+  private String gcpProjectId;
+
+  @CommandLine.Option(names = "--dataset-id", required = true, description = "BigQuery dataset id.")
+  private String bigQueryDatasetId;
+
   @CommandLine.Mixin WorkspaceOverride workspaceOption;
   @CommandLine.Mixin Format formatOption;
 
@@ -34,8 +42,8 @@ public class BqDataset extends BaseCommand {
     CreateBqDatasetParams.Builder createParams =
         new CreateBqDatasetParams.Builder()
             .resourceFields(createResourceParams.build())
-            .projectId(bigQueryIds.getGcpProjectId())
-            .datasetId(bigQueryIds.getBigQueryDatasetId());
+            .projectId(gcpProjectId)
+            .datasetId(bigQueryDatasetId);
 
     bio.terra.cli.businessobject.resource.BqDataset createdResource =
         bio.terra.cli.businessobject.resource.BqDataset.addReferenced(createParams.build());
