+++ src/main/java/bio/terra/cli/serialization/userfacing/UFResource.java
@@ -1,7 +1,6 @@
 package bio.terra.cli.serialization.userfacing;
 
 import bio.terra.cli.businessobject.Resource;
-import bio.terra.cli.serialization.userfacing.resource.UFBqDataTable;
 import bio.terra.cli.serialization.userfacing.resource.UFBqDataset;
 import bio.terra.cli.serialization.userfacing.resource.UFGcpNotebook;
 import bio.terra.cli.serialization.userfacing.resource.UFGcsBucket;
@@ -33,7 +32,6 @@ import java.util.UUID;
     property = "resourceType")
 @JsonSubTypes({
   @Type(value = UFGcpNotebook.class, name = "AI_NOTEBOOK"),
-  @Type(value = UFBqDataTable.class, name = "BQ_DATA_TABLE"),
   @Type(value = UFBqDataset.class, name = "BQ_DATASET"),
   @Type(value = UFGcsBucket.class, name = "GCS_BUCKET")
 })
